chatgpt: clean up temporary files during uploadRepoFiles
We noticed our gerrit instance running out of disk space and found a
series of json files generated under `/tmp`. The structure looked like:
/tmp/<repository_name><hash>.json

We notice these temp files would get generated upon a CR change-merged
event:
"<timestamp & user> INFO  com.googlesource.gerrit.plugins.aicodereview
.listener.GerritListener : Processing event: com.google.gerrit.server
.events.ChangeMergedEvent@commit-hash"

The timestamp of this event would always match the timestamp of the
json file created under `tmp`. These files also only get created for
CRs merged in a repo where ai-code-review is created.

We pin-pointed the issue to createTempFileWithContent in this plugin.
There is a deleteOnExit() clause for this, but it does not seem to
delete the files in all situations.

Some of these files are quite large, as they contain the whole content
of a CR.

This change is a proposed fix for deleting a file in a finally clause
in case it has not been removed yet and still exists.

Change-Id: I1fb62f95c20b509cc756a0702d112e5eb44cc20a